split up index file
