+++ docs/api_getLinkStatus.js.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Home</title>
+    <title>JSDoc: Source: api/getLinkStatus.js</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,32 +17,53 @@
 
 <div id="main">
 
-    <h1 class="page-title">Home</h1>
+    <h1 class="page-title">Source: api/getLinkStatus.js</h1>
 
     
 
 
 
     
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import {
+    assert,
+    getAbstractSigner,
+    getParamsFromLink
+} from '../common/index.js'
+import {
+    getContract
+} from './index.js'
 
+/**
+ * Checks if a link has been claimed
+ *
+ * @param {Object} options - An object containing the signer and link to check
+ * @returns {Object} - An object containing whether the link has been claimed and the deposit
+ */
+export async function getLinkStatus({ signer, link }) {
+    /* checks if a link has been claimed */
+    assert(signer, 'signer arg is required')
+    assert(link, 'link arg is required')
 
-    <h3> </h3>
+    signer = await getAbstractSigner(signer)
 
+    const params = getParamsFromLink(link)
+    const chainId = params.chainId
+    const contractVersion = params.contractVersion
+    const depositIdx = params.depositIdx
+    const contract = await getContract(chainId, signer, contractVersion)
+    const deposit = await contract.deposits(depositIdx)
 
-
-
-
-
-
-
-
-
-    
-
-
-
-
-
+    // if the deposit is claimed, the pubKey20 will be 0x000....
+    if (deposit.pubKey20 == '0x0000000000000000000000000000000000000000') {
+        return { claimed: true, deposit }
+    }
+    return { claimed: false, deposit }
+}
+</code></pre>
+        </article>
+    </section>
