+++ docs/claimLink.js.html
@@ -1,0 +1,141 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: claimLink.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: claimLink.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import { ethers } from 'ethersv5'
+import {
+	assert,
+	generateKeysFromString,
+	getAbstractSigner,
+	getParamsFromLink,
+	signAddress,
+	solidityHashAddress,
+	solidityHashBytesEIP191,
+} from '../common/index.js'
+import {
+	getContract,
+	setFeeOptions
+} from './index.js'
+
+/**
+ * Claims the contents of a link
+ *
+ * @param {Object} options - An object containing the options to use for claiming the link
+ * @param {Object} options.signer - The signer to use for claiming
+ * @param {string} options.link - The link to claim
+ * @param {string} [options.recipient=null] - The address to claim the link to. Defaults to the signer's address if not provided
+ * @param {boolean} [options.verbose=false] - Whether or not to print verbose output
+ * @returns {Object} - The transaction receipt
+ */
+export async function claimLink({
+	signer,
+	link,
+	recipient = null,
+	verbose = false,
+	maxFeePerGas = null,
+	maxPriorityFeePerGas = null,
+	gasLimit = null,
+	eip1559 = true,
+}) {
+	// claims the contents of a link
+	assert(signer, 'signer arg is required')
+	assert(link, 'link arg is required')
+
+	signer = await getAbstractSigner(signer)
+
+	const params = getParamsFromLink(link)
+	const chainId = params.chainId
+	const contractVersion = params.contractVersion
+	const depositIdx = params.depositIdx
+	const password = params.password
+	if (recipient == null) {
+		recipient = await signer.getAddress()
+		verbose &amp;&amp; console.log('recipient not provided, using signer address: ', recipient)
+	}
+	const keys = generateKeysFromString(password) // deterministically generate keys from password
+	const contract = await getContract(chainId, signer, contractVersion)
+
+	// cryptography
+	var addressHash = solidityHashAddress(recipient)
+	var addressHashBinary = ethers.utils.arrayify(addressHash) // v5
+	verbose &amp;&amp; console.log('addressHash: ', addressHash, ' addressHashBinary: ', addressHashBinary)
+	var addressHashEIP191 = solidityHashBytesEIP191(addressHashBinary)
+	var signature = await signAddress(recipient, keys.privateKey) // sign with link keys
+
+	if (verbose) {
+		// print the params
+		console.log('params: ', params)
+		console.log('addressHash: ', addressHash)
+		console.log('addressHashEIP191: ', addressHashEIP191)
+		console.log('signature: ', signature)
+	}
+
+	// Prepare transaction options
+	let txOptions = {}
+	txOptions = await setFeeOptions({
+		txOptions,
+		provider: signer.provider,
+		eip1559,
+		maxFeePerGas,
+		maxPriorityFeePerGas,
+		gasLimit,
+		verbose,
+	})
+
+	const claimParams = [depositIdx, recipient, addressHashEIP191, signature]
+	verbose &amp;&amp; console.log('claimParams: ', claimParams)
+
+	// withdraw the deposit
+	const tx = await contract.withdrawDeposit(...claimParams, txOptions)
+	console.log('submitted tx: ', tx.hash, ' now waiting for receipt...')
+	const txReceipt = await tx.wait()
+
+	return txReceipt
+}
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#approveSpendERC20">approveSpendERC20</a></li><li><a href="global.html#claimLink">claimLink</a></li><li><a href="global.html#claimLinkGasless">claimLinkGasless</a></li><li><a href="global.html#claimLinkSender">claimLinkSender</a></li><li><a href="global.html#createLink">createLink</a></li><li><a href="global.html#getAllDepositsForSigner">getAllDepositsForSigner</a></li><li><a href="global.html#getContract">getContract</a></li><li><a href="global.html#getDefaultProvider">getDefaultProvider</a></li><li><a href="global.html#getLinkDetails">getLinkDetails</a></li><li><a href="global.html#getLinkStatus">getLinkStatus</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Tue Aug 22 2023 22:37:10 GMT+0200 (Mitteleuropäische Sommerzeit)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
