+++ docs/claimLinkGasless.js.html
@@ -1,0 +1,109 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: claimLinkGasless.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: claimLinkGasless.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import { createClaimPayload } from './index.js'
+
+/**
+ * Claims a link through the Peanut API
+ *
+ * @param {string} link - The link to claim
+ * @param {string} recipientAddress - The recipient address to claim the link to
+ * @param {string} apiKey -The API key to use for the Peanut API
+ * @param {string} url - The URL to use for the Peanut API (default is 'https://api.peanut.to/claim')
+ * @returns {Object} - The data returned from the API call
+ */
+export async function claimLinkGasless(
+	link,
+	recipientAddress,
+	apiKey,
+	verbose = false,
+	url = 'https://api.peanut.to/claim'
+) {
+	console.log('claiming link through Peanut API...')
+	verbose &amp;&amp; console.log('link: ', link, ' recipientAddress: ', recipientAddress, ' apiKey: ', apiKey, ' url: ', url)
+	const payload = await createClaimPayload(link, recipientAddress)
+	verbose &amp;&amp; console.log('payload: ', payload)
+	//  url = "https://api.peanut.to/claim";
+	if (url == 'local') {
+		console.log('using local api')
+		url = 'http://127.0.0.1:5001/claim'
+	}
+
+	const headers = {
+		'Content-Type': 'application/json',
+	}
+
+	const body = {
+		address: payload.recipientAddress,
+		address_hash: payload.addressHash,
+		signature: payload.signature,
+		idx: payload.idx,
+		chain: payload.chainId,
+		version: payload.contractVersion,
+		api_key: apiKey,
+	}
+
+	// if axios error, return the error message
+
+	const response = await fetch(url, {
+		method: 'POST',
+		headers: headers,
+		body: JSON.stringify(body),
+	})
+
+	if (!response.ok) {
+		const error = await response.text()
+		throw new Error(error)
+	} else {
+		const data = await response.json()
+		return data
+	}
+}
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#approveSpendERC20">approveSpendERC20</a></li><li><a href="global.html#claimLink">claimLink</a></li><li><a href="global.html#claimLinkGasless">claimLinkGasless</a></li><li><a href="global.html#claimLinkSender">claimLinkSender</a></li><li><a href="global.html#createLink">createLink</a></li><li><a href="global.html#getAllDepositsForSigner">getAllDepositsForSigner</a></li><li><a href="global.html#getContract">getContract</a></li><li><a href="global.html#getDefaultProvider">getDefaultProvider</a></li><li><a href="global.html#getLinkDetails">getLinkDetails</a></li><li><a href="global.html#getLinkStatus">getLinkStatus</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Tue Aug 22 2023 22:37:10 GMT+0200 (Mitteleuropäische Sommerzeit)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
