+++ docs/claimLinkSender.js.html
@@ -1,0 +1,122 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: claimLinkSender.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: claimLinkSender.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import { ethers } from 'ethersv5'
+import {
+    assert,
+    generateKeysFromString,
+    getAbstractSigner,
+    getParamsFromLink,
+    signAddress,
+    solidityHashAddress,
+    solidityHashBytesEIP191,
+} from '../common/index.js'
+import {
+    getContract
+} from './index.js'
+
+/**
+ * Claims the contents of a link as a sender. Can only be used if a link has not been claimed in a set time period.
+ * (24 hours). Only works with links created with v4 of the contract. More gas efficient than claimLink.
+ *
+ * @param {Object} options - An object containing the options to use for claiming the link
+ * @param {Object} options.signer - The signer to use for claiming
+ * @param {string} options.link - The link to claim
+ * @param {boolean} [options.verbose=false] - Whether or not to print verbose output
+ * @returns {Object} - The transaction receipt
+ */
+export async function claimLinkSender({ signer, link, verbose = false }) {
+    // raise error, not implemented yet
+    throw new Error('Not implemented yet')
+    assert(signer, 'signer arg is required')
+    assert(link, 'link arg is required')
+
+    signer = await getAbstractSigner(signer)
+
+    const params = getParamsFromLink(link)
+    const chainId = params.chainId
+    const contractVersion = params.contractVersion
+    const depositIdx = params.depositIdx
+    const password = params.password
+    if (recipient == null) {
+        recipient = await signer.getAddress()
+
+        verbose &amp;&amp; console.log('recipient not provided, using signer address: ', recipient)
+    }
+    const keys = generateKeysFromString(password) // deterministically generate keys from password
+    const contract = await getContract(chainId, signer, contractVersion)
+
+    // cryptography
+    var addressHash = solidityHashAddress(recipient)
+    // var addressHashBinary = ethers.getBytes(addressHash); // v6
+    var addressHashBinary = ethers.utils.arrayify(addressHash) // v5
+    verbose &amp;&amp; console.log('addressHash: ', addressHash, ' addressHashBinary: ', addressHashBinary)
+    var addressHashEIP191 = solidityHashBytesEIP191(addressHashBinary)
+    var signature = await signAddress(recipient, keys.privateKey) // sign with link keys
+
+    if (verbose) {
+        // print the params
+        console.log('params: ', params)
+        console.log('addressHash: ', addressHash)
+        console.log('addressHashEIP191: ', addressHashEIP191)
+        console.log('signature: ', signature)
+    }
+
+    // TODO: use createClaimPayload instead
+
+    // withdraw the deposit
+    // address hash is hash(prefix + hash(address))
+    const tx = await contract.withdrawDeposit(depositIdx, recipient, addressHashEIP191, signature)
+    console.log('submitted tx: ', tx.hash, ' now waiting for receipt...')
+    const txReceipt = await tx.wait()
+
+    return txReceipt
+}
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#approveSpendERC20">approveSpendERC20</a></li><li><a href="global.html#claimLink">claimLink</a></li><li><a href="global.html#claimLinkGasless">claimLinkGasless</a></li><li><a href="global.html#claimLinkSender">claimLinkSender</a></li><li><a href="global.html#createLink">createLink</a></li><li><a href="global.html#getAllDepositsForSigner">getAllDepositsForSigner</a></li><li><a href="global.html#getContract">getContract</a></li><li><a href="global.html#getDefaultProvider">getDefaultProvider</a></li><li><a href="global.html#getLinkDetails">getLinkDetails</a></li><li><a href="global.html#getLinkStatus">getLinkStatus</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Tue Aug 22 2023 22:37:10 GMT+0200 (Mitteleuropäische Sommerzeit)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
