+++ docs/common_util.js.html
@@ -1,5 +1,33 @@
-import { ethers } from 'ethersv5' // v5
-import { CHAIN_MAP } from './data.js'
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: common/util.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: common/util.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import { ethers } from 'ethersv5'
+import { CHAIN_MAP } from './index.js'
 
 export function assert(condition, message) {
 	if (!condition) {
@@ -12,7 +40,7 @@ export function assert(condition, message) {
  */
 export function greeting() {
 	console.log(
-		'🥜 Hello & thanks for using the Peanut SDK! If you run into any issues, dm @hugomont on telegram or hop on the Peanut Protocol discord'
+		'🥜 Hello &amp; thanks for using the Peanut SDK! If you run into any issues, dm @hugomont on telegram or hop on the Peanut Protocol discord'
 	)
 }
 
@@ -115,7 +143,7 @@ export async function signAddress(string, privateKey) {
 export function getRandomString(length) {
 	const chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
 	let result_str = ''
-	for (let i = 0; i < length; i++) {
+	for (let i = 0; i &lt; length; i++) {
 		result_str += chars[Math.floor(Math.random() * chars.length)]
 	}
 	return result_str
@@ -200,10 +228,10 @@ export function getLinkFromParams(
 ) {
 	/* returns a link from the given parameters */
 
-	const link = baseUrl + '#?c=' + chainId + '&v=' + contractVersion + '&i=' + depositIdx + '&p=' + password
+	const link = baseUrl + '#?c=' + chainId + '&amp;v=' + contractVersion + '&amp;i=' + depositIdx + '&amp;p=' + password
 
 	if (trackId != '') {
-		return link + '&t=' + trackId
+		return link + '&amp;t=' + trackId
 	}
 	return link
 }
@@ -252,7 +280,7 @@ export function getDepositIdxs(txReceipt, chainId, contractAddress) {
 	const logs = txReceipt.logs
 	var depositIdxs = []
 	// loop through all the logs and extract the deposit index from each
-	for (var i = 0; i < logs.length; i++) {
+	for (var i = 0; i &lt; logs.length; i++) {
 		// check if the log was emitted by our contract
 		if (logs[i].address.toLowerCase() === contractAddress.toLowerCase()) {
 			if (chainId == 137) {
@@ -264,3 +292,26 @@ export function getDepositIdxs(txReceipt, chainId, contractAddress) {
 	}
 	return depositIdxs
 }
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#approveSpendERC20">approveSpendERC20</a></li><li><a href="global.html#claimLink">claimLink</a></li><li><a href="global.html#claimLinkGasless">claimLinkGasless</a></li><li><a href="global.html#claimLinkSender">claimLinkSender</a></li><li><a href="global.html#createLink">createLink</a></li><li><a href="global.html#generateKeysFromString">generateKeysFromString</a></li><li><a href="global.html#getAllDepositsForSigner">getAllDepositsForSigner</a></li><li><a href="global.html#getContract">getContract</a></li><li><a href="global.html#getDefaultProvider">getDefaultProvider</a></li><li><a href="global.html#getDepositIdx">getDepositIdx</a></li><li><a href="global.html#getDepositIdxs">getDepositIdxs</a></li><li><a href="global.html#getLinkDetails">getLinkDetails</a></li><li><a href="global.html#getLinkFromParams">getLinkFromParams</a></li><li><a href="global.html#getLinkStatus">getLinkStatus</a></li><li><a href="global.html#getParamsFromLink">getParamsFromLink</a></li><li><a href="global.html#getParamsFromPageURL">getParamsFromPageURL</a></li><li><a href="global.html#getRandomString">getRandomString</a></li><li><a href="global.html#greeting">greeting</a></li><li><a href="global.html#hash_string">hash_string</a></li><li><a href="global.html#signAddress">signAddress</a></li><li><a href="global.html#signMessageWithPrivatekey">signMessageWithPrivatekey</a></li><li><a href="global.html#solidityHashAddress">solidityHashAddress</a></li><li><a href="global.html#solidityHashBytesEIP191">solidityHashBytesEIP191</a></li><li><a href="global.html#verifySignature">verifySignature</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Tue Aug 22 2023 22:37:51 GMT+0200 (Mitteleuropäische Sommerzeit)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
