+++ docs/getAllDeposittsForSigner.js.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: getAllDeposittsForSigner.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: getAllDeposittsForSigner.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import {
+    DEFAULT_CONTRACT_VERSION
+} from '../common/data.js'
+import {
+    getContract
+} from './index.js'
+
+/**
+ * Gets all deposits for a given signer and chainId.
+ *
+ */
+export async function getAllDepositsForSigner({
+    signer,
+    chainId,
+    contractVersion = DEFAULT_CONTRACT_VERSION,
+    verbose = false,
+}) {
+    const contract = await getContract(chainId, signer, contractVersion)
+    let deposits
+    if (contractVersion == 'v3') {
+        // throw warning if using v3
+        console.warn('WARNING: This function is not efficient for v3 contracts. Not recommended to use.')
+        const depositCount = await contract.getDepositCount()
+        deposits = []
+        for (let i = 0; i &lt; depositCount; i++) {
+            verbose &amp;&amp; console.log('fetching deposit: ', i)
+            let deposit = await contract.deposits(i)
+            deposits.push(deposit)
+        }
+    } else {
+        // v4: we now have getAllDeposits available
+        const address = await signer.getAddress()
+        // const allDeposits = await contract.getAllDeposits();
+        deposits = await contract.getAllDepositsForAddress(address)
+    }
+    return deposits
+}
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#approveSpendERC20">approveSpendERC20</a></li><li><a href="global.html#claimLink">claimLink</a></li><li><a href="global.html#claimLinkGasless">claimLinkGasless</a></li><li><a href="global.html#claimLinkSender">claimLinkSender</a></li><li><a href="global.html#createLink">createLink</a></li><li><a href="global.html#getAllDepositsForSigner">getAllDepositsForSigner</a></li><li><a href="global.html#getContract">getContract</a></li><li><a href="global.html#getDefaultProvider">getDefaultProvider</a></li><li><a href="global.html#getLinkDetails">getLinkDetails</a></li><li><a href="global.html#getLinkStatus">getLinkStatus</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Tue Aug 22 2023 22:37:10 GMT+0200 (Mitteleuropäische Sommerzeit)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
