+++ docs/getLinkDetails.js.html
@@ -1,0 +1,131 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: getLinkDetails.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: getLinkDetails.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>import { ethers } from 'ethersv5'
+import {
+    TOKEN_DETAILS,
+    assert,
+    getParamsFromLink
+} from '../common/index.js'
+import { getContract } from './web3.js'
+
+/**
+ * Gets the details of a Link: what token it is, how much it holds, etc.
+ *
+ * @param {Object} signerOrProvider - The signer or provider to use
+ * @param {string} link - The link to get the details of
+ * @param {boolean} verbose - Whether or not to print verbose output
+ * @returns {Object} - An object containing the link details
+ */
+export async function getLinkDetails(signerOrProvider, link, verbose = false) {
+    /**
+     * Gets the details of a Link: what token it is, how much it holds, etc.
+     */
+    verbose &amp;&amp; console.log('getLinkDetails called with link: ', link)
+    assert(signerOrProvider, 'signerOrProvider arg is required')
+    assert(link, 'link arg is required')
+
+    const params = getParamsFromLink(link)
+    const chainId = params.chainId
+    const contractVersion = params.contractVersion
+    const depositIdx = params.depositIdx
+    const password = params.password
+    const contract = await getContract(chainId, signerOrProvider, contractVersion)
+
+    const deposit = await contract.deposits(depositIdx)
+    var tokenAddress = deposit.tokenAddress
+    verbose &amp;&amp; console.log('fetched deposit: ', deposit)
+
+    const tokenType = deposit.contractType
+    verbose &amp;&amp; console.log('tokenType: ', tokenType, typeof tokenType)
+
+    if (tokenType == 0) {
+        // native token, set zero address
+        // TODO: is this a potential footgun or no? Why is matic 0xeeeeee....? Is this a problem?
+        verbose &amp;&amp; console.log('tokenType is 0, setting tokenAddress to zero address')
+        tokenAddress = ethers.constants.AddressZero
+    }
+    verbose &amp;&amp; console.log('deposit: ', deposit)
+
+    // Retrieve the token's details from the tokenDetails.json file
+    verbose &amp;&amp; console.log('finding token details for token with address: ', tokenAddress, ' on chain: ', chainId)
+    // Find the correct chain details using chainId
+    console.log('chainId: ', chainId)
+    const chainDetails = TOKEN_DETAILS.find((chain) => chain.chainId === String(chainId))
+    if (!chainDetails) {
+        throw new Error('Chain details not found')
+    }
+
+    // Find the token within the tokens array of the chain
+    const tokenDetails = chainDetails.tokens.find((token) => token.address.toLowerCase() === tokenAddress.toLowerCase())
+    if (!tokenDetails) {
+        throw new Error('Token details not found')
+    }
+
+    // Format the token amount
+    const tokenAmount = ethers.utils.formatUnits(deposit.amount, tokenDetails.decimals)
+
+    // TODO: Fetch token price using API
+
+    return {
+        link: link,
+        chainId: chainId,
+        depositIndex: depositIdx,
+        contractVersion: contractVersion,
+        password: password,
+        tokenType: deposit.contractType,
+        tokenAddress: deposit.tokenAddress,
+        tokenSymbol: tokenDetails.symbol,
+        tokenName: tokenDetails.name,
+        tokenAmount: tokenAmount,
+        // tokenPrice: tokenPrice
+    }
+}
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Global</h3><ul><li><a href="global.html#approveSpendERC20">approveSpendERC20</a></li><li><a href="global.html#claimLink">claimLink</a></li><li><a href="global.html#claimLinkGasless">claimLinkGasless</a></li><li><a href="global.html#claimLinkSender">claimLinkSender</a></li><li><a href="global.html#createLink">createLink</a></li><li><a href="global.html#getAllDepositsForSigner">getAllDepositsForSigner</a></li><li><a href="global.html#getContract">getContract</a></li><li><a href="global.html#getDefaultProvider">getDefaultProvider</a></li><li><a href="global.html#getLinkDetails">getLinkDetails</a></li><li><a href="global.html#getLinkStatus">getLinkStatus</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Tue Aug 22 2023 22:37:10 GMT+0200 (Mitteleuropäische Sommerzeit)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
