+++ src/common/core.js
