+++ src/common/signer.js
@@ -1,0 +1,6 @@
+// implements abstract signer class
+
+export async function getAbstractSigner(signer, verbose = true) {
+	// TODO: create abstract signer class that is compatible with ethers v5, v6, viem, web3js
+	return signer
+}
